Update tests related to results
